+++ commitlint.config.js
@@ -16,11 +16,4 @@ limitations under the License.
 
 */
 
-export * from './inline-form'
-export * from './inline-field'
-export * from './inline-field-text'
-export * from './inline-field-textarea'
-export * from './inline-wysiwyg'
-export * from './inline-image-field'
-export * from './blocks'
-export * from './styles'
+module.exports = { extends: ['@commitlint/config-conventional'] }