+++ packages/react-tinacms-editor/src/state/plugins/inline-menu/Menu.tsx
@@ -71,8 +71,6 @@ const LinkControl = markControl({
   defaultAttrs: {
     href: '',
     title: '',
-    editing: 'editing',
-    creating: 'creating',
   },
   noMix: ['code'],
 })