+++ packages/gatsby-plugin-tinacms/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.28](https://github.com/tinacms/tinacms/compare/gatsby-plugin-tinacms@0.1.27...gatsby-plugin-tinacms@0.1.28) (2020-03-16)
+
+**Note:** Version bump only for package gatsby-plugin-tinacms
+
+
+
+
+
 ## [0.1.27](https://github.com/tinacms/tinacms/compare/gatsby-plugin-tinacms@0.1.27-alpha.1...gatsby-plugin-tinacms@0.1.27) (2020-03-09)
 
 **Note:** Version bump only for package gatsby-plugin-tinacms