+++ packages/gatsby-tinacms-remark/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.7](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-remark@0.7.6...gatsby-tinacms-remark@0.7.7) (2020-03-16)
+
+**Note:** Version bump only for package gatsby-tinacms-remark
+
+
+
+
+
 ## [0.7.6](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-remark@0.7.6-alpha.1...gatsby-tinacms-remark@0.7.6) (2020-03-09)
 
 **Note:** Version bump only for package gatsby-tinacms-remark